+++ R/board-server.R
@@ -35,7 +35,7 @@ board_server.board <- function(id, x, plugins = board_plugins(x),
                                callback_location = c("end", "start"),
                                ...) {
 
-  reload_meta <- finalize_reload("reload")
+  reload_meta <- finalize_reload(staged_board_key(session_url_query()))
 
   plugins <- as_plugins(plugins)
 
@@ -238,53 +238,22 @@ board_server.board <- function(id, x, plugins = board_plugins(x),
 
       if (not_null(board_refresh)) {
 
-        reload_pending <- reactiveVal(NULL)
+        observeEvent(
+          board_refresh(),
+          {
+            val <- board_refresh()
 
-        observeEvent(board_refresh(), reload_pending(Sys.time()))
-
-        observe({
-
-          pending <- reload_pending()
-          req(pending)
-
-          if (is_reloading("reload")) {
-
-            if (difftime(Sys.time(), pending, units = "secs") >= 60) {
-              log_warn("stale reload state, clearing")
-              finalize_reload("reload")
-              reload_pending(NULL)
-              session$close()
-              return()
+            if (is_board(val)) {
+              board <- val
+              meta <- NULL
+            } else {
+              board <- val$board
+              meta <- val$meta
             }
 
-            notify(
-              "Reload in progress.",
-              duration = NULL,
-              id = session$token
-            )
-            invalidateLater(1000)
-            return()
+            stage_reload_handoff(board, meta, session)
           }
-
-          removeNotification(session$token)
-          reload_pending(NULL)
-
-          val <- isolate(board_refresh())
-
-          if (is_board(val)) {
-            board <- val
-            meta <- NULL
-          } else {
-            board <- val$board
-            meta <- val$meta
-          }
-
-          log_debug("refreshing board")
-          update_serve_obj("reload", board, meta = meta)
-
-          log_debug("reloading session")
-          session$reload()
-        })
+        )
       }
 
       call_plugin_server(